requirements.txt
playwright>=1.40.0
PyYAML>=6.0
rich>=13.0.0
keyring>=24.0.0
